switched from size_t to unsigned

+++ src/ccdoubles-cplx-matrices.c
@@ -37,20 +37,20 @@
  ** ----------------------------------------------------------------- */
 
 void
-ccdoubles_cplx_matrix_clear (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_clear (unsigned nrows, unsigned ncols,
 			     double complex * restrict matrix)
 {
   ccdoubles_cplx_vector_clear(nrows * ncols, matrix);
 }
 void
-ccdoubles_cplx_matrix_set (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_set (unsigned nrows, unsigned ncols,
 			   double complex * restrict matrix,
 			   double complex value)
 {
   ccdoubles_cplx_vector_set(nrows * ncols, matrix, value);
 }
 void
-ccdoubles_cplx_matrix_copy (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_copy (unsigned nrows, unsigned ncols,
 			    double complex * restrict dst,
 			    double complex * restrict src)
 {
@@ -60,35 +60,35 @@ ccdoubles_cplx_matrix_copy (size_t nrows, size_t ncols,
 /* ------------------------------------------------------------------ */
 
 void
-ccdoubles_cplx_matrix_real (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_real (unsigned nrows, unsigned ncols,
 			    double * restrict result,
 			    double complex * restrict operand)
 {
   ccdoubles_cplx_vector_real (nrows * ncols, result, operand);
 }
 void
-ccdoubles_cplx_matrix_imag (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_imag (unsigned nrows, unsigned ncols,
 			    double * restrict result,
 			    double complex * restrict operand)
 {
   ccdoubles_cplx_vector_imag (nrows * ncols, result, operand);
 }
 void
-ccdoubles_cplx_matrix_magnitude (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_magnitude (unsigned nrows, unsigned ncols,
 				 double * restrict result,
 				 double complex * restrict operand)
 {
   ccdoubles_cplx_vector_magnitude (nrows * ncols, result, operand);
 }
 void
-ccdoubles_cplx_matrix_angle (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_angle (unsigned nrows, unsigned ncols,
 			     double * restrict result,
 			     double complex * restrict operand)
 {
   ccdoubles_cplx_vector_angle (nrows * ncols, result, operand);
 }
 void
-ccdoubles_cplx_matrix_conj (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_conj (unsigned nrows, unsigned ncols,
 			    double complex * restrict result,
 			    double complex * restrict operand)
 {
@@ -98,7 +98,7 @@ ccdoubles_cplx_matrix_conj (size_t nrows, size_t ncols,
 /* ------------------------------------------------------------------ */
 
 void
-ccdoubles_cplx_matrix_from_rect (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_from_rect (unsigned nrows, unsigned ncols,
 				 double complex * restrict result,
 				 double * restrict real,
 				 double * restrict imag)
@@ -106,7 +106,7 @@ ccdoubles_cplx_matrix_from_rect (size_t nrows, size_t ncols,
   ccdoubles_cplx_vector_from_rect (nrows * ncols, result, real, imag);
 }
 void
-ccdoubles_cplx_matrix_from_polar (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_from_polar (unsigned nrows, unsigned ncols,
 				  double complex * restrict result,
 				  double * restrict magnitude,
 				  double * restrict angle)
@@ -120,7 +120,7 @@ ccdoubles_cplx_matrix_from_polar (size_t nrows, size_t ncols,
  ** ----------------------------------------------------------------- */
 
 void
-ccdoubles_cplx_matrix_add (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_add (unsigned nrows, unsigned ncols,
 			   double complex * restrict result,
 			   double complex * restrict operand1,
 			   double complex * restrict operand2)
@@ -128,7 +128,7 @@ ccdoubles_cplx_matrix_add (size_t nrows, size_t ncols,
   ccdoubles_cplx_vector_add(nrows * ncols, result, operand1, operand2);
 }
 void
-ccdoubles_cplx_matrix_sub (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_sub (unsigned nrows, unsigned ncols,
 			   double complex * restrict result,
 			   double complex * restrict operand1,
 			   double complex * restrict operand2)
@@ -136,7 +136,7 @@ ccdoubles_cplx_matrix_sub (size_t nrows, size_t ncols,
   ccdoubles_cplx_vector_sub(nrows * ncols, result, operand1, operand2);
 }
 void
-ccdoubles_cplx_matrix_mul (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_mul (unsigned nrows, unsigned ncols,
 			   double complex * restrict result,
 			   double complex * restrict operand1,
 			   double complex * restrict operand2)
@@ -144,7 +144,7 @@ ccdoubles_cplx_matrix_mul (size_t nrows, size_t ncols,
   ccdoubles_cplx_vector_mul(nrows * ncols, result, operand1, operand2);
 }
 void
-ccdoubles_cplx_matrix_div (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_div (unsigned nrows, unsigned ncols,
 			   double complex * restrict result,
 			   double complex * restrict operand1,
 			   double complex * restrict operand2)
@@ -152,7 +152,7 @@ ccdoubles_cplx_matrix_div (size_t nrows, size_t ncols,
   ccdoubles_cplx_vector_div(nrows * ncols, result, operand1, operand2);
 }
 void
-ccdoubles_cplx_matrix_neg (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_neg (unsigned nrows, unsigned ncols,
 			   double complex * restrict result,
 			   double complex * restrict operand)
 {
@@ -165,7 +165,7 @@ ccdoubles_cplx_matrix_neg (size_t nrows, size_t ncols,
  ** ----------------------------------------------------------------- */
 
 void
-ccdoubles_cplx_matrix_scalar_mul (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_scalar_mul (unsigned nrows, unsigned ncols,
 				  double complex * restrict result,
 				  double complex lambda,
 				  double complex * restrict operand)
@@ -173,7 +173,7 @@ ccdoubles_cplx_matrix_scalar_mul (size_t nrows, size_t ncols,
   ccdoubles_cplx_vector_scalar_mul(nrows * ncols, result, lambda, operand);
 }
 void
-ccdoubles_cplx_matrix_linear_combination (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_linear_combination (unsigned nrows, unsigned ncols,
 					  double complex * restrict result,
 					  double complex alpha,
 					  double complex * restrict operand1,
@@ -184,7 +184,7 @@ ccdoubles_cplx_matrix_linear_combination (size_t nrows, size_t ncols,
 					   result, alpha, operand1, beta, operand2);
 }
 void
-ccdoubles_cplx_matrix_transpose (size_t operand_nrows, size_t operand_ncols,
+ccdoubles_cplx_matrix_transpose (unsigned operand_nrows, unsigned operand_ncols,
 				 double complex * restrict result,
 				 double complex * restrict operand)
 /* To call this function we are meant to do:
@@ -194,14 +194,14 @@ ccdoubles_cplx_matrix_transpose (size_t operand_nrows, size_t operand_ncols,
  *    ccdoubles_cplx_matrix_transpose (2, 3, &R[0][0], &O[0][0]);
  */
 {
-  for (size_t i=0; i<operand_nrows; ++i) {
-    for (size_t j=0; j<operand_ncols; ++j) {
+  for (unsigned i=0; i<operand_nrows; ++i) {
+    for (unsigned j=0; j<operand_ncols; ++j) {
       result[j * operand_nrows + i] = operand[i * operand_ncols + j];
     }
   }
 }
 void
-ccdoubles_cplx_matrix_conjugate_transpose (size_t operand_nrows, size_t operand_ncols,
+ccdoubles_cplx_matrix_conjugate_transpose (unsigned operand_nrows, unsigned operand_ncols,
 					   double complex * restrict result,
 					   double complex * restrict operand)
 /* To call this function we are meant to do:
@@ -211,8 +211,8 @@ ccdoubles_cplx_matrix_conjugate_transpose (size_t operand_nrows, size_t operand_
  *    ccdoubles_cplx_matrix_conjugate_transpose (2, 3, &R[0][0], &O[0][0]);
  */
 {
-  for (size_t i=0; i<operand_nrows; ++i) {
-    for (size_t j=0; j<operand_ncols; ++j) {
+  for (unsigned i=0; i<operand_nrows; ++i) {
+    for (unsigned j=0; j<operand_ncols; ++j) {
       if (i != j) {
 	result[j * operand_nrows + i] = conj(operand[i * operand_ncols + j]);
       } else {
@@ -222,9 +222,9 @@ ccdoubles_cplx_matrix_conjugate_transpose (size_t operand_nrows, size_t operand_
   }
 }
 void
-ccdoubles_cplx_matrix_rowcol_mul (size_t result_nrows,
-				  size_t operand_n,
-				  size_t result_ncols,
+ccdoubles_cplx_matrix_rowcol_mul (unsigned result_nrows,
+				  unsigned operand_n,
+				  unsigned result_ncols,
 				  double complex * restrict result,
 				  double complex * restrict operand1,
 				  double complex * restrict operand2)
@@ -238,11 +238,11 @@ ccdoubles_cplx_matrix_rowcol_mul (size_t result_nrows,
  *                                     &R[0][0], &O1[0][0], &O2[0][0]);
  */
 {
-  for (size_t i=0; i<result_nrows; ++i) {
-    for (size_t j=0; j<result_ncols; ++j) {
+  for (unsigned i=0; i<result_nrows; ++i) {
+    for (unsigned j=0; j<result_ncols; ++j) {
       double complex *	R = &(result[i * result_ncols + j]);
       *R = 0.0;
-      for (size_t k=0; k<operand_n; ++k) {
+      for (unsigned k=0; k<operand_n; ++k) {
 	*R += operand1[i * operand_n + k] * operand2[k * result_ncols + j];
       }
     }
@@ -255,35 +255,35 @@ ccdoubles_cplx_matrix_rowcol_mul (size_t result_nrows,
  ** ----------------------------------------------------------------- */
 
 void
-ccdoubles_cplx_matrix_exp (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_exp (unsigned nrows, unsigned ncols,
 			   double complex * restrict result,
 			   double complex * restrict operand)
 {
   ccdoubles_cplx_vector_exp (nrows * ncols, result, operand);
 }
 void
-ccdoubles_cplx_matrix_log (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_log (unsigned nrows, unsigned ncols,
 			   double complex * restrict result,
 			   double complex * restrict operand)
 {
   ccdoubles_cplx_vector_log (nrows * ncols, result, operand);
 }
 void
-ccdoubles_cplx_matrix_log10 (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_log10 (unsigned nrows, unsigned ncols,
 			     double complex * restrict result,
 			     double complex * restrict operand)
 {
   ccdoubles_cplx_vector_log10 (nrows * ncols, result, operand);
 }
 void
-ccdoubles_cplx_matrix_sqrt (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_sqrt (unsigned nrows, unsigned ncols,
 			   double complex * restrict result,
 			   double complex * restrict operand)
 {
   ccdoubles_cplx_vector_sqrt (nrows * ncols, result, operand);
 }
 void
-ccdoubles_cplx_matrix_pow (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_pow (unsigned nrows, unsigned ncols,
 			   double complex * restrict result,
 			   double complex * restrict operand1,
 			   double complex * restrict operand2)
@@ -297,21 +297,21 @@ ccdoubles_cplx_matrix_pow (size_t nrows, size_t ncols,
  ** ----------------------------------------------------------------- */
 
 void
-ccdoubles_cplx_matrix_sin (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_sin (unsigned nrows, unsigned ncols,
 			   double complex * restrict result,
 			   double complex * restrict operand)
 {
   ccdoubles_cplx_vector_sin (nrows * ncols, result, operand);
 }
 void
-ccdoubles_cplx_matrix_cos (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_cos (unsigned nrows, unsigned ncols,
 			   double complex * restrict result,
 			   double complex * restrict operand)
 {
   ccdoubles_cplx_vector_cos (nrows * ncols, result, operand);
 }
 void
-ccdoubles_cplx_matrix_tan (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_tan (unsigned nrows, unsigned ncols,
 			   double complex * restrict result,
 			   double complex * restrict operand)
 {
@@ -324,21 +324,21 @@ ccdoubles_cplx_matrix_tan (size_t nrows, size_t ncols,
  ** ----------------------------------------------------------------- */
 
 void
-ccdoubles_cplx_matrix_asin (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_asin (unsigned nrows, unsigned ncols,
 			    double complex * restrict result,
 			    double complex * restrict operand)
 {
   ccdoubles_cplx_vector_asin (nrows * ncols, result, operand);
 }
 void
-ccdoubles_cplx_matrix_acos (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_acos (unsigned nrows, unsigned ncols,
 			    double complex * restrict result,
 			    double complex * restrict operand)
 {
   ccdoubles_cplx_vector_acos (nrows * ncols, result, operand);
 }
 void
-ccdoubles_cplx_matrix_atan (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_atan (unsigned nrows, unsigned ncols,
 			    double complex * restrict result,
 			    double complex * restrict operand)
 {
@@ -351,21 +351,21 @@ ccdoubles_cplx_matrix_atan (size_t nrows, size_t ncols,
  ** ----------------------------------------------------------------- */
 
 void
-ccdoubles_cplx_matrix_sinh (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_sinh (unsigned nrows, unsigned ncols,
 			    double complex * restrict result,
 			    double complex * restrict operand)
 {
   ccdoubles_cplx_vector_sinh (nrows * ncols, result, operand);
 }
 void
-ccdoubles_cplx_matrix_cosh (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_cosh (unsigned nrows, unsigned ncols,
 			    double complex * restrict result,
 			    double complex * restrict operand)
 {
   ccdoubles_cplx_vector_cosh (nrows * ncols, result, operand);
 }
 void
-ccdoubles_cplx_matrix_tanh (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_tanh (unsigned nrows, unsigned ncols,
 			    double complex * restrict result,
 			    double complex * restrict operand)
 {
@@ -378,21 +378,21 @@ ccdoubles_cplx_matrix_tanh (size_t nrows, size_t ncols,
  ** ----------------------------------------------------------------- */
 
 void
-ccdoubles_cplx_matrix_asinh (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_asinh (unsigned nrows, unsigned ncols,
 			     double complex * restrict result,
 			     double complex * restrict operand)
 {
   ccdoubles_cplx_vector_asinh (nrows * ncols, result, operand);
 }
 void
-ccdoubles_cplx_matrix_acosh (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_acosh (unsigned nrows, unsigned ncols,
 			     double complex * restrict result,
 			     double complex * restrict operand)
 {
   ccdoubles_cplx_vector_acosh (nrows * ncols, result, operand);
 }
 void
-ccdoubles_cplx_matrix_atanh (size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_atanh (unsigned nrows, unsigned ncols,
 			     double complex * restrict result,
 			     double complex * restrict operand)
 {
@@ -406,18 +406,18 @@ ccdoubles_cplx_matrix_atanh (size_t nrows, size_t ncols,
 
 void
 ccdoubles_cplx_matrix_print_display (FILE * f, const char * name,
-				     size_t nrows, size_t ncols,
+				     unsigned nrows, unsigned ncols,
 				     double complex * operand)
 {
-  size_t	i, j;
-  fprintf(f, "Row-major matrix %s (dimension %ld x %ld) (displayed in row-major order):\n",
+  unsigned	i, j;
+  fprintf(f, "Row-major matrix %s (dimension %u x %u) (displayed in row-major order):\n",
 	  name, nrows, ncols);
   for (i=0; i<nrows; ++i) {
     j = 0;
-    fprintf(f, "| (%ld,%ld) %+lf%-+lfi ", 1+i, 1+j,
+    fprintf(f, "| (%u,%u) %+lf%-+lfi ", 1+i, 1+j,
 	    creal(operand[i * ncols + j]), cimag(operand[i * ncols + j]));
     for (++j; j<ncols; ++j) {
-      fprintf(f, "; (%ld,%ld) %+lf%-+lfi ", 1+i, 1+j,
+      fprintf(f, "; (%u,%u) %+lf%-+lfi ", 1+i, 1+j,
 	      creal(operand[i * ncols + j]), cimag(operand[i * ncols + j]));
     }
     fprintf(f, "|\n");
@@ -425,10 +425,10 @@ ccdoubles_cplx_matrix_print_display (FILE * f, const char * name,
   fprintf(f, "\n");
 }
 void
-ccdoubles_cplx_matrix_print_brackets (FILE * f, size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_print_brackets (FILE * f, unsigned nrows, unsigned ncols,
 				      double complex * operand)
 {
-  size_t	i, j;
+  unsigned	i, j;
   fprintf(f, "[[%+lf%-+lfi", creal(operand[0]), cimag(operand[0]));
   for (j=1; j<ncols; ++j) {
     fprintf(f, " %+lf%-+lfi", creal(operand[j]), cimag(operand[j]));
@@ -445,10 +445,10 @@ ccdoubles_cplx_matrix_print_brackets (FILE * f, size_t nrows, size_t ncols,
   fprintf(f, "]\n");
 }
 void
-ccdoubles_cplx_matrix_print_braces (FILE * f, size_t nrows, size_t ncols,
+ccdoubles_cplx_matrix_print_braces (FILE * f, unsigned nrows, unsigned ncols,
 				    double complex * operand)
 {
-  size_t	i, j;
+  unsigned	i, j;
   fprintf(f, "{{%+lf%-+lfi", creal(operand[0]), cimag(operand[0]));
   for (j=1; j<ncols; ++j) {
     fprintf(f, ", %+lf%-+lfi", creal(operand[j]), cimag(operand[j]));
