+++ src/ccdoubles-int-vectors.c
@@ -37,24 +37,24 @@
  ** ----------------------------------------------------------------- */
 
 void
-ccdoubles_int_vector_clear (size_t nslots, int * restrict vector)
+ccdoubles_int_vector_clear (unsigned nslots, int * restrict vector)
 {
   memset(vector, 0, sizeof(int) * nslots);
 }
 void
-ccdoubles_int_vector_set (size_t nslots, int * restrict vector, int value)
+ccdoubles_int_vector_set (unsigned nslots, int * restrict vector, int value)
 {
-  for (size_t i=0; i<nslots; ++i) {
+  for (unsigned i=0; i<nslots; ++i) {
     vector[i] = value;
   }
 }
 void
-ccdoubles_int_vector_copy (size_t nslots, int * restrict dst, int * restrict src)
+ccdoubles_int_vector_copy (unsigned nslots, int * restrict dst, int * restrict src)
 {
   if (1) {
     memcpy(dst, src, sizeof(int) * nslots);
   } else {
-    for (size_t i=0; i<nslots; ++i) {
+    for (unsigned i=0; i<nslots; ++i) {
       dst[i] = src[i];
     }
   }
@@ -67,30 +67,30 @@ ccdoubles_int_vector_copy (size_t nslots, int * restrict dst, int * restrict src
 
 void
 ccdoubles_int_vector_print_display (FILE * f, const char * name,
-				    size_t nslots,
+				    unsigned nslots,
 				    int * operand)
 {
-  fprintf(f, "Vector %s (dimension %ld):\n", name, nslots);
+  fprintf(f, "Vector %s (dimension %u):\n", name, nslots);
   fprintf(f, "| (1) %+d ", operand[0]);
-  for (size_t i=1; i<nslots; ++i) {
-    fprintf(f, "; (%ld) %+d ", 1+i, operand[i]);
+  for (unsigned i=1; i<nslots; ++i) {
+    fprintf(f, "; (%u) %+d ", 1+i, operand[i]);
   }
   fprintf(f, "|\n");
 }
 void
-ccdoubles_int_vector_print_brackets (FILE * f, size_t nslots, int * operand)
+ccdoubles_int_vector_print_brackets (FILE * f, unsigned nslots, int * operand)
 {
   fprintf(f, "[%+d", operand[0]);
-  for (size_t i=1; i<nslots; ++i) {
+  for (unsigned i=1; i<nslots; ++i) {
     fprintf(f, " %+d", operand[i]);
   }
   fprintf(f, "]\n");
 }
 void
-ccdoubles_int_vector_print_braces (FILE * f, size_t nslots, int * operand)
+ccdoubles_int_vector_print_braces (FILE * f, unsigned nslots, int * operand)
 {
   fprintf(f, "{%+d", operand[0]);
-  for (size_t i=1; i<nslots; ++i) {
+  for (unsigned i=1; i<nslots; ++i) {
     fprintf(f, ", %+d", operand[i]);
   }
   fprintf(f, "}\n");
